pyproject.toml
[tool.poetry]
name = "auth-sprint-1"
version = "0.1.0"
description = "Authentication module for our online cimena"
authors = [
    "Artem Samoilov <samoylovartem07@gmail.com>",
    "Imran Dzhumabaev <safeisnotanoption@gmail.com>",
    "Sergey Timoshenkov <serkoe2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask-Security-Too = "^5.1.2"
flask-jwt-extended = "^4.4.4"
Flask-SQLAlchemy = "^3.0.3"
Flask-Migrate = "^4.0.4"
redis = "^4.4.4"
pydantic = "^1.10.7"
orjson = "^3.8.9"
alembic = "^1.10.2"
dependency-injector = "^4.41.0"
psycopg2-binary = "^2.9.6"
requests = "^2.28.2"
gevent = "^22.10.2"
typer = "^0.7.0"
python-dotenv = "^1.0.0"
apispec = "^6.3.0"
apispec-webframeworks = "^0.5.2"
flask-cors = "^3.0.10"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.2"


[tool.isort]
profile = "black"
known_first_party = "auth_api, src"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
lines_between_types = 1


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
